moved upload process function from controller to service for RIM
