+++ HIRS_AttestationCAPortal/src/main/java/hirs/attestationca/portal/page/controllers/ReferenceManifestPageController.java
@@ -1,17 +1,15 @@
 package hirs.attestationca.portal.page.controllers;
 
+import hirs.attestationca.persist.dto.PageMessages;
 import hirs.attestationca.persist.entity.userdefined.DataTablesColumn;
 import hirs.attestationca.persist.entity.userdefined.DownloadFile;
 import hirs.attestationca.persist.entity.userdefined.FilteredRecordsList;
 import hirs.attestationca.persist.entity.userdefined.ReferenceManifest;
-import hirs.attestationca.persist.entity.userdefined.rim.BaseReferenceManifest;
-import hirs.attestationca.persist.entity.userdefined.rim.SupportReferenceManifest;
 import hirs.attestationca.persist.service.ReferenceManifestPageService;
 import hirs.attestationca.portal.datatables.DataTableInput;
 import hirs.attestationca.portal.datatables.DataTableResponse;
 import hirs.attestationca.portal.datatables.Order;
 import hirs.attestationca.portal.page.Page;
-import hirs.attestationca.portal.page.PageMessages;
 import hirs.attestationca.portal.page.params.NoPageParams;
 import hirs.attestationca.portal.page.utils.ControllerPagesUtils;
 import jakarta.servlet.http.HttpServletResponse;
@@ -42,7 +40,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
-import java.util.regex.Pattern;
 import java.util.zip.ZipOutputStream;
 
 /**
@@ -52,9 +49,6 @@ import java.util.zip.ZipOutputStream;
 @RequestMapping("/HIRS_AttestationCAPortal/portal/reference-manifests")
 @Log4j2
 public class ReferenceManifestPageController extends PageController<NoPageParams> {
-
-    private static final String BASE_RIM_FILE_PATTERN = "(\\S+(\\.(?i)swidtag)$)";
-    private static final String SUPPORT_RIM_FILE_PATTERN = "(\\S+(\\.(?i)(rimpcr|rimel|bin|log))$)";
 
     private final ReferenceManifestPageService referenceManifestPageService;
 
@@ -140,71 +134,9 @@ public class ReferenceManifestPageController extends PageController<NoPageParams
     protected RedirectView uploadRIMs(@RequestParam("file") final MultipartFile[] files,
                                       final RedirectAttributes redirectAttributes)
             throws URISyntaxException {
+
         Map<String, Object> model = new HashMap<>();
-        PageMessages messages = new PageMessages();
-        List<String> successMessages = new ArrayList<>();
-        List<String> errorMessages = new ArrayList<>();
-
-        final Pattern baseRimPattern = Pattern.compile(BASE_RIM_FILE_PATTERN);
-        final Pattern supportRimPattern = Pattern.compile(SUPPORT_RIM_FILE_PATTERN);
-
-        List<BaseReferenceManifest> baseRims = new ArrayList<>();
-        List<SupportReferenceManifest> supportRims = new ArrayList<>();
-
-        log.info("Uploading {} RIM files", files.length);
-
-        for (MultipartFile file : files) {
-            String fileName = file.getOriginalFilename();
-
-            if (fileName == null || fileName.isEmpty()) {
-                log.warn("File with empty or null name skipped");
-                continue;  // Skip processing this file
-            }
-
-            final boolean isBaseRim = baseRimPattern.matcher(fileName).matches();
-            final boolean isSupportRim = !isBaseRim && supportRimPattern.matcher(fileName).matches();
-
-            if (isBaseRim) {
-                final BaseReferenceManifest baseReferenceManifest =
-                        referenceManifestPageService.parseBaseRIM(errorMessages, file);
-                baseRims.add(baseReferenceManifest);
-                messages.addErrorMessages(errorMessages);
-                if (baseReferenceManifest != null) {
-                    log.info("Uploaded base RIM with manufacturer {} and model {}.",
-                            baseReferenceManifest.getPlatformManufacturer(),
-                            baseReferenceManifest.getPlatformModel());
-                } else {
-                    log.info("Failed to parse Base RIM file {}", fileName);
-                }
-            } else if (isSupportRim) {
-                final SupportReferenceManifest supportReferenceManifest =
-                        referenceManifestPageService.parseSupportRIM(errorMessages, file);
-                supportRims.add(supportReferenceManifest);
-                messages.addErrorMessages(errorMessages);
-                if (supportReferenceManifest != null) {
-                    log.info("Uploaded support RIM with manufacturer {} and model {}.",
-                            supportReferenceManifest.getPlatformManufacturer(),
-                            supportReferenceManifest.getPlatformModel());
-                    String associatedBaseRim = supportReferenceManifest.getAssociatedRim() != null
-                            ? supportReferenceManifest.getAssociatedRim().toString() : "not found";
-                    log.info("Associated base RIM {}", associatedBaseRim);
-                } else {
-                    log.info("Failed to parse support RIM file {}", fileName);
-                }
-            } else {
-                String errorString = "The file extension of " + fileName + " was not recognized."
-                        + " Base RIMs support the extension \".swidtag\", and support RIMs support "
-                        + "\".rimpcr\", \".rimel\", \".bin\", and \".log\". "
-                        + "Please verify your upload and retry.";
-                log.error("File extension in {} not recognized as base or support RIM.", fileName);
-                errorMessages.add(errorString);
-            }
-        }
-
-        this.referenceManifestPageService.storeRIMS(successMessages, errorMessages, baseRims, supportRims);
-
-        messages.addSuccessMessages(successMessages);
-        messages.addErrorMessages(errorMessages);
+        PageMessages messages = referenceManifestPageService.processUploads(files);
 
         model.put(MESSAGES_ATTRIBUTE, messages);
         return redirectTo(Page.REFERENCE_MANIFESTS, new NoPageParams(), model, redirectAttributes);
