+++ HIRS_AttestationCAPortal/src/test/java/hirs/attestationca/portal/page/controllers/EndorsementKeyCredentialsPageControllerTest.java
@@ -4,7 +4,7 @@ import hirs.attestationca.persist.entity.manager.EndorsementCertificateRepositor
 import hirs.attestationca.persist.entity.userdefined.Certificate;
 import hirs.attestationca.persist.entity.userdefined.certificate.EndorsementCredential;
 import hirs.attestationca.portal.page.PageControllerTest;
-import hirs.attestationca.portal.page.PageMessages;
+import hirs.attestationca.persist.dto.PageMessages;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
